Fix bug involving plugin config default values
If config.plugins.sassLint is provided, then the default values
for both "file" and "options" will be lost.

For example, if config.plugins.sassLint specifies "options"
but doesn't specify "file", then the plugin won't supply
a default value for "file".

An implementation which avoids this problem can be seen here:

https://github.com/brunch/coffeelint-brunch/blob/2.0.0/index.js#L13-L19
https://github.com/brunch/eslint-brunch/blob/v3.11.1/index.js#L43-L46
https://github.com/brunch/jshint-brunch/blob/2.0.0/index.js#L28-L34